Dictionary
The code creates a dictionary 'fruits' that maps various fruit names to their respective calorie counts. The user is then prompted to enter a fruit name, which is converted to title case using the title() method. The code then loops through the keys in the 'fruits' dictionary, and for each key that is found in the user's input (using the 'in' keyword), it prints the corresponding value (i.e., the number of calories).

So essentially, the code is looking up the number of calories for a fruit entered by the user and printing it to the console. Note that the 'title()' method is used to standardize the user input, since the keys in the 'fruits' dictionary are all in title case.