Guessing game
This code prompts the user to enter a positive integer as the level. It generates a random number within the range of 1 to the level. 

The user is then prompted to make a guess. If the guess is smaller than the random number, it displays "Too small!". If the guess is larger than the random number, it displays "Too large!". If the guess is equal to the random number, it displays "Just right!" and terminates the loop.

The program continues to ask for guesses until the user correctly guesses the random number.