Number Guessing Game
**Title: Random Number Guessing Game**

**Description:**

This Python code is a concise random number guessing game. It prompts the user to enter an upper limit for the range and validates the input. After generating a random number within the specified range, the program allows the user to guess the number. It provides feedback on each attempt until the correct number is guessed. The game concludes by displaying the number of guesses it took to win.